+++ frontend/js/config.js
@@ -1,3 +1,3 @@
-const API_URL = '%%BACKEND_URL%%';
+const API_URL = 'http://localhost:8080';
 
 export { API_URL };